# Optimize Enumerable.count/1 implementation for File.Stream