--- a/lib/elixir/lib/file/stream.ex
+++ b/lib/elixir/lib/file/stream.ex
@@ -67,12 +67,12 @@ defmodule File.Stream do
   end
 
   defimpl Enumerable do
-    def reduce(%{path: path, modes: modes, line_or_bytes: line_or_bytes, raw: raw}, acc, fun) do
-      modes = for mode <- modes, not mode in [:write, :append], do: mode
+    @read_ahead_size 64 * 1024
 
+    def reduce(%{path: path, modes: modes, line_or_bytes: line_or_bytes, raw: raw}, acc, fun) do
       start_fun =
         fn ->
-          case :file.open(path, modes) do
+          case :file.open(path, read_modes(modes)) do
             {:ok, device}    -> device
             {:error, reason} ->
               raise File.Error, reason: reason, action: "stream", path: path
@@ -88,12 +88,51 @@ defmodule File.Stream do
       Stream.resource(start_fun, next_fun, &:file.close/1).(acc, fun)
     end
 
-    def count(_stream) do
-      {:error, __MODULE__}
+    def count(%{path: path, modes: modes, line_or_bytes: :line} = stream) do
+      pattern = :binary.compile_pattern("\n")
+      counter = &count_lines(&1, path, pattern, read_function(stream), 0)
+
+      case File.open(path, read_modes(modes), counter) do
+        {:ok, count} ->
+          {:ok, count}
+        {:error, reason} ->
+          raise File.Error, reason: reason, action: "stream", path: path
+      end
+    end
+
+    def count(%{path: path, line_or_bytes: bytes}) do
+      case File.stat(path) do
+        {:ok, %{size: 0}} ->
+          {:error, __MODULE__}
+        {:ok, %{size: size}} ->
+          {:ok, div(size, bytes) + if(rem(size, bytes) == 0, do: 0, else: 1)}
+        {:error, reason} ->
+          raise File.Error, reason: reason, action: "stream", path: path
+      end
     end
 
     def member?(_stream, _term) do
       {:error, __MODULE__}
     end
+
+    defp read_modes(modes) do
+      for mode <- modes, not mode in [:write, :append], do: mode
+    end
+
+    defp count_lines(device, path, pattern, read, count) do
+      case read.(device) do
+        data when is_binary(data) ->
+          count_lines(device, path, pattern, read, count + count_lines(data, pattern))
+        :eof ->
+          count
+        {:error, reason} ->
+          raise File.Error, reason: reason, action: "stream", path: path
+      end
+    end
+
+    defp count_lines(data, pattern), do: length(:binary.matches(data, pattern))
+
+    defp read_function(%{raw: true}), do: &IO.binread(&1, @read_ahead_size)
+    defp read_function(%{raw: false}), do: &IO.read(&1, @read_ahead_size)
   end
 end